# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Outdoor Event Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{id}/registrations:
    get:
      operationId: listEventRegistrations
      tags: [registrations]
      summary: List registrations for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
...
  /registrations/{id}:
    delete:
      operationId: cancelRegistration
      tags: [registrations]
      summary: Cancel a registration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cancelled
    patch:
      operationId: updateRegistrationStatus
      tags: [registrations]
      summary: Update registration status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationStatusUpdate"
      responses:
        "200":
          description: Updated registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "404":
          description: Not found
...
  /public/register/{token}:
    get:
      operationId: getEventByToken
      tags: [public]
      summary: Get event details by registration token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event with sponsors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventWithSponsors"
        "404":
          description: Not found
    post:
      operationId: registerForEvent
      tags: [public]
      summary: Submit a registration for an event
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "400":
          description: Bad request
  /public/photos/{token}:
    get:
      operationId: getEventByPhotoToken
      tags: [public]
      summary: Get event details and photos by photo token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event with photos and sponsors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventWithPhotosAndSponsors"
        "404":
          description: Not found
    post:
      operationId: uploadPhotoByToken
      tags: [public]
      summary: Submit a photo for an event via photo token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoInput"
      responses:
        "201":
          description: Photo submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        "404":
          description: Not found
...
components:
  schemas:
...
    Event:
      type: object
      required: [id, title, category, date, location, capacity, registrationCount, status, createdAt, registrationToken, photoToken, pointsValue]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [cycling, hiking, summer-night, walking]
        eventTypeId:
          type: ["integer", "null"]
        date:
          type: string
          format: date-time
        location:
          type: string
        capacity:
          type: integer
        registrationCount:
          type: integer
        status:
          type: string
          enum: [upcoming, ongoing, completed, cancelled]
        difficulty:
          type: ["string", "null"]
          enum: [easy, moderate, challenging, null]
        distance:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        meetingPoint:
          type: ["string", "null"]
        guidelines:
          type: ["string", "null"]
        pointsValue:
          type: integer
        registrationToken:
          type: string
        photoToken:
          type: string
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    RegistrationStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, waitlist, cancelled]
    Registration:
      type: object
      required: [id, eventId, name, email, status, registeredAt, referralCount]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        name:
          type: string
        fullNameAr:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: ["string", "null"]
        phoneCountryCode:
          type: string
        nationality:
          type: ["string", "null"]
        nationalityOther:
          type: ["string", "null"]
        hasMedicalConditions:
          type: boolean
        medicalDetails:
          type: ["string", "null"]
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactPhone:
          type: ["string", "null"]
        waiverAcceptedAt:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [pending, confirmed, waitlist, cancelled]
        referralToken:
          type: ["string", "null"]
        referredByRegistrationId:
          type: ["integer", "null"]
        referralCount:
          type: integer
        registeredAt:
          type: string
          format: date-time
    RegistrationInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 1
        fullNameAr:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        phoneCountryCode:
          type: string
        nationality:
          type: string
        nationalityOther:
          type: string
        hasMedicalConditions:
          type: boolean
        medicalDetails:
          type: string
        emergencyContactName:
          type: string
        emergencyContactPhone:
          type: string
        waiverAcceptedAt:
          type: string
          format: date-time
        refToken:
          type: string
...
    Photo:
      type: object
      required: [id, eventId, uploaderName, objectPath, uploadedAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        uploaderName:
          type: string
        caption:
          type: ["string", "null"]
        objectPath:
          type: string
        uploadedAt:
          type: string
          format: date-time
    PhotoInput:
      type: object
      required: [uploaderName, objectPath]
      properties:
        uploaderName:
          type: string
          minLength: 1
        caption:
          type: string
        objectPath:
          type: string
          minLength: 1
...
    EventWithSponsors:
      type: object
      required: [event, sponsors]
      properties:
        event:
          $ref: "#/components/schemas/Event"
        sponsors:
          type: array
          items:
            $ref: "#/components/schemas/Sponsor"
    EventWithPhotosAndSponsors:
      type: object
      required: [event, photos, sponsors]
      properties:
        event:
          $ref: "#/components/schemas/Event"
        photos:
          type: array
          items:
            $ref: "#/components/schemas/Photo"
        sponsors:
          type: array
          items:
            $ref: "#/components/schemas/Sponsor"
    Sponsor:
      type: object
      required: [id, name, type, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [cafe, restaurant, camping, hotel, gym, shop, other]
        website:
          type: ["string", "null"]
        instagram:
          type: ["string", "null"]
        facebook:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        discountCode:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...