# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Outdoor Event Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationInput {
  /** @minLength 1 */
  name: string;
  fullNameAr?: string;
  email: string;
  phone?: string;
  phoneCountryCode?: string;
  nationality?: string;
  nationalityOther?: string;
  hasMedicalConditions?: boolean;
  medicalDetails?: string;
  emergencyContactName?: string;
  emergencyContactPhone?: string;
  waiverAcceptedAt?: Date;
  refToken?: string;
}