# lib/api-zod/src/generated/types/sponsor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Outdoor Event Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { SponsorType } from './sponsorType';

export interface Sponsor {
  id: number;
  name: string;
  type: SponsorType;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  facebook?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  discountCode?: string | null;
  createdAt: Date;
}